feat: RTTI 추가 완료
메모리 해제 시 Array 관리 확인 필요
Scene Load Save 로직 때 오브젝트 관리 체크.

+++ GTLEngine/Source/CoreUObject/Components/PrimitiveComponent.cpp
@@ -49,7 +49,7 @@ bool UPrimitiveComponent::IsRayIntersect(FRay ray, float hitDistance, FVector& h
     bool result = false;
     //BYTE* pbPositions = reinterpret_cast<BYTE*>(vertices.data());
 
-    int nPrimitives = vertices.size() / 3;
+    int nPrimitives = static_cast<int>(vertices.size() / 3);
 
     float nearHitDistancePow = FLT_MAX;
     FVector nearestHitPoint = FVector::Zero();
